requirements.txt
flask
python-dotenv
google-generativeai
grpcio
edge-tts
aiohttp
speechrecognition
pydub
gunicorn